backend/requirements.txt
Flask
flask-cors
python-dotenv
openai
google-generativeai
requests
pytest